move print dataset test to new file
